+++ src/world/Map/Area/AreaStorage.hpp
@@ -13,7 +13,7 @@ namespace MapManagement
 {
     namespace AreaManagement
     {
-        typedef std::map<uint16, uint32> AreaFlagByAreaID;
+        typedef std::map<uint32, uint32> AreaFlagByAreaID;
         typedef std::map<uint32, uint32> AreaFlagByMapID;
 
         // Temporary 
@@ -57,7 +57,7 @@ namespace MapManagement
                 static bool IsOutdoor(uint32 mapId, float x, float y, float z);
                 static bool IsOutdoorWMO(uint32 mogpFlags, int32 /*adtId*/, int32 /*rootId*/, int32 /*groupId*/, DBC::Structures::WMOAreaTableEntry const* wmoEntry, DBC::Structures::AreaTableEntry const* atEntry);
                 
-                static const uint16 GetFlagByPosition(uint16 area_flag_without_adt_id, bool have_area_info, uint32 mogp_flags, int32 adt_id, int32 root_id, int32 group_id, uint32 map_id, float x, float y, float z, bool* _out_is_outdoors = nullptr);
+                static const uint32 GetFlagByPosition(uint32 area_flag_without_adt_id, bool have_area_info, uint32 mogp_flags, int32 adt_id, int32 root_id, int32 group_id, uint32 map_id, float x, float y, float z, bool* _out_is_outdoors = nullptr);
         };
     } // </ AreaManagementNamespace>
 } // </ MapManagementNamespace>
