+++ src/world/Map/MapMgr.h
@@ -83,7 +83,7 @@ class SERVER_DECL MapMgr : public CellHandler <MapCell>, public EventableObject,
 
         CObjectFactory ObjectFactory;
 
-        const uint16 GetAreaFlag(float x, float y, float z, bool *is_outdoors = nullptr);
+        const uint32 GetAreaFlag(float x, float y, float z, bool *is_outdoors = nullptr);
 
         /// This will be done in regular way soon
         std::set<MapCell*> m_forcedcells;
